refactor: update namespaces to use Config\Schema for configuration classes
