+++ src/Module/Config/Schema/Embed/Software.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Internal\DLoad\Module\Common\Config\Embed;
+namespace Internal\DLoad\Module\Config\Schema\Embed;
 
-use Internal\DLoad\Module\Common\Internal\Attribute\XPathEmbed;
 use Internal\DLoad\Module\Common\Internal\Attribute\XPath;
+use Internal\DLoad\Module\Common\Internal\Attribute\XPathEmbed;
 use Internal\DLoad\Module\Common\Internal\Attribute\XPathEmbedList;
 
 /**
