+++ src/Module/Config/Schema/GitHub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Internal\DLoad\Module\Common\Config;
+namespace Internal\DLoad\Module\Config\Schema;
 
 use Internal\DLoad\Module\Common\Internal\Attribute\Env;
