+++ src/Module/Downloader/Task/DownloadTask.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Downloader\Task;
 
-use Internal\DLoad\Module\Common\Config\Embed\Software;
+use Internal\DLoad\Module\Config\Schema\Embed\Software;
 use Internal\DLoad\Module\Downloader\Progress;
 use React\Promise\PromiseInterface;
