+++ src/Module/Repository/Internal/GitHub/Factory.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Repository\Internal\GitHub;
 
-use Internal\DLoad\Module\Common\Config\Embed\Repository as RepositoryConfig;
-use Internal\DLoad\Module\Common\Config\GitHub as GitHubConfig;
+use Internal\DLoad\Module\Config\Schema\Embed\Repository as RepositoryConfig;
+use Internal\DLoad\Module\Config\Schema\GitHub as GitHubConfig;
 use Internal\DLoad\Module\Repository\RepositoryFactory;
 use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
